refactor: extract `useFormState` hooks
