+++ apps/calendar/src/components/popup/dateSelector.tsx
@@ -4,10 +4,10 @@ import { useEffect, useRef } from 'preact/hooks';
 
 import DatePicker, { DateRangePicker } from 'tui-date-picker';
 
-import { FormStateActionType, FormStateDispatcher } from '@src/components/popup/eventFormPopup';
 import { PopupSection } from '@src/components/popup/popupSection';
 import { useStore } from '@src/contexts/calendarStore';
 import { cls } from '@src/helpers/css';
+import { FormStateActionType, FormStateDispatcher } from '@src/hooks/popup/formState';
 import { optionsSelector } from '@src/selectors';
 import TZDate from '@src/time/date';
