+++ apps/calendar/src/components/popup/titleInputBox.tsx
@@ -1,8 +1,8 @@
 import { FunctionComponent, h } from 'preact';
 
-import { FormStateActionType, FormStateDispatcher } from '@src/components/popup/eventFormPopup';
 import { PopupSection } from '@src/components/popup/popupSection';
 import { cls } from '@src/helpers/css';
+import { FormStateActionType, FormStateDispatcher } from '@src/hooks/popup/formState';
 
 interface Props {
   isPrivate?: boolean;
