+++ apps/calendar/src/hooks/popup/formState.ts
@@ -1,0 +1,40 @@
+import { useReducer } from 'preact/hooks';
+
+import { EventModelData, EventState } from '@t/events';
+
+export enum FormStateActionType {
+  setCalendarId = 'setCalendarId',
+  setPrivate = 'setPrivate',
+  setAllday = 'setAllday',
+  setState = 'setState',
+}
+
+type FormStateAction =
+  | { type: FormStateActionType.setCalendarId; calendarId: string }
+  | { type: FormStateActionType.setPrivate; isPrivate: boolean }
+  | { type: FormStateActionType.setAllday; isAllday: boolean }
+  | { type: FormStateActionType.setState; state: EventState };
+
+export type FormStateDispatcher = (action: FormStateAction) => void;
+
+function formStateReducer(state: EventModelData, action: FormStateAction): EventModelData {
+  switch (action.type) {
+    case FormStateActionType.setCalendarId:
+      return { ...state, calendarId: action.calendarId };
+    case FormStateActionType.setPrivate:
+      return { ...state, isPrivate: action.isPrivate };
+    case FormStateActionType.setAllday:
+      return { ...state, isAllday: action.isAllday };
+    case FormStateActionType.setState:
+      return { ...state, state: action.state };
+
+    default:
+      return state;
+  }
+}
+
+export function useFormState(initialState: EventModelData) {
+  const [formState, formStateDispatch] = useReducer(formStateReducer, initialState);
+
+  return { formState, formStateDispatch };
+}
